+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - update deltachat-node and deltachat/jsonrpc-client to `v1.109.0`
 
 ### Fixed
-- better error handling when messages fail to load from db
+- better error handling when messages fail to load from db in messagelist and gallery
+
 
 ### Removed
 - remove unused direct dependency on babel: `@babel/core`, `@babel/preset-env` and `@babel/preset-react`